+++ launch/launch/legacy/main.py
@@ -16,9 +16,9 @@ import argparse
 import os
 import sys
 
-from launch import LaunchDescriptor
-from launch.launcher import DefaultLauncher
-from launch.loader import load_launch_file
+from launch.legacy import LaunchDescriptor
+from launch.legacy.launcher import DefaultLauncher
+from launch.legacy.loader import load_launch_file
 
 
 def file_exists(filename):
